working BM25F TF-IDF versions

+++ neo4j-ir-procedures/src/main/java/improvedSearch/BM25F.java
@@ -151,8 +151,8 @@ public class BM25F {
 
         terms.forEach((k,v)->{
             if(termsStartsWith(v, qkw.getStem())) {
-                int[] tfField = (int[]) occurrence.get(k);
-                int tempOccurrence = tfField[termPosition.get(k).get(CURRENT_TERM)];
+                double[] tfField = (double[]) occurrence.get(k);
+                double tempOccurrence = tfField[termPosition.get(k).get(CURRENT_TERM)];
                 double tfFieldScore = tfField(length.get(k), (Double) fieldAvgLength.get(k), tempOccurrence, k);
 
                 sum.updateAndGet(v1 -> (v1 + boost.get(k) * tfFieldScore));
@@ -161,7 +161,7 @@ public class BM25F {
         return sum.get();
     }
 
-    public double tfField(int length, double avgL, int occurrence, String fieldName){
+    public double tfField(int length, double avgL, double occurrence, String fieldName){
         double bField = b.get(fieldName);
         return occurrence/(1+bField*((length/avgL)-1));
     }
